+++ extensions/scaleway/src/iam/pages/Policies/components/index.ts
@@ -1,2 +1,2 @@
-export { Policy } from './Policy'
 export { DropDownOrderBy } from './DropDownOrderBy'
+export { Policy } from './Policy'